% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.1st}
\alias{cheb.1st}
\title{Generating Chebyshev Polynomials of first kind}
\usage{
cheb.1st(x.axis, n)
}
\arguments{
\item{x.axis}{beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb.t Chebyshev-Polynome Erster Art (Vektor)
}
\description{
cheb.1st erzeugt Chebyshev Polynome erster Art aus beliebiger X-Achse
}

